// packages/backend/src/fplapi/typings/interfaces/entry-picks-response.interface.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface EntryPicksResponse {
  activeChip?: null;
  automaticSubs?: {
    entry?: number;
    elementIn?: number;
    elementOut?: number;
    event?: number;
    [k: string]: unknown;
  }[];
  entryHistory?: {
    event?: number;
    points?: number;
    totalPoints?: number;
    rank?: number;
    rankSort?: number;
    overallRank?: number;
    bank?: number;
    value?: number;
    eventTransfers?: number;
    eventTransfersCost?: number;
    pointsOnBench?: number;
    [k: string]: unknown;
  };
  picks?: {
    element?: number;
    position?: number;
    multiplier?: number;
    isCaptain?: boolean;
    isViceCaptain?: boolean;
    [k: string]: unknown;
  }[];
  [k: string]: unknown;
}